refactor: extract reusable Collective Map canvas
Moves all Leaflet rendering logic (MapContainer, TileLayer, marker switch, icon builders, popup components, jitter, popupCss) out of map-view.tsx into a new CollectiveMapCanvas component. MapView becomes a thin card wrapper that renders the canvas at height=480 so /peta is visually identical. The height prop (number | string, default 480) allows future fullscreen usage in /maps without touching /peta. peta-client.tsx is unchanged — its dynamic ssr:false import covers the canvas transitively.
